refactor utility func to kafka component
